tests(lantern): refactor to DevTools test convention
