+++ .eslintignore
@@ -19,3 +19,5 @@ third-party/**
 **/*.d.cts
 
 page-functions-test-case*out*.js
+# TODO(15841): remove when importing Lantern from npm
+core/lib/lantern/**/*.test.js
