+++ core/lib/lantern/metrics/SpeedIndex.test.js
@@ -4,8 +4,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import * as Lantern from '../../../../lib/lantern/lantern.js';
-import {readJson} from '../../../test-utils.js';
+import * as Lantern from '../lantern.js';
+import {readJson} from '../../../test/test-utils.js';
 import {getComputationDataFromFixture} from './MetricTestUtils.js';
 
 const {SpeedIndex, FirstContentfulPaint} = Lantern.Metrics;
