+++ core/lib/lantern/simulation/Simulator.test.js
@@ -6,9 +6,9 @@
 
 import assert from 'assert/strict';
 
-import * as Lantern from '../../../../lib/lantern/lantern.js';
-import {readJson} from '../../../test-utils.js';
-import {runTraceEngine} from '../metrics/MetricTestUtils.js';
+import * as Lantern from '../lantern.js';
+import {readJson} from '../../../test/test-utils.js';
+import {runTraceEngine} from '../../../test/lib/lantern/metrics/MetricTestUtils.js';
 
 const {NetworkNode, CPUNode} = Lantern;
 const {Simulator, DNSCache} = Lantern.Simulation;
