+++ tsconfig.json
@@ -66,13 +66,15 @@
     "core/test/lib/emulation-test.js",
     "core/test/lib/i18n/i18n-test.js",
     "core/test/lib/icons-test.js",
-    "core/test/lib/lantern/base-node-test.js",
-    "core/test/lib/lantern/metrics/*",
-    "core/test/lib/lantern/page-dependency-graph-test.js",
-    "core/test/lib/lantern/simulation/connection-pool-test.js",
-    "core/test/lib/lantern/simulation/dns-cache-test.js",
-    "core/test/lib/lantern/simulation/network-analyzer-test.js",
-    "core/test/lib/lantern/simulation/simulator-test.js",
+    // TODO(15841): remove when importing Lantern from npm
+    "core/lib/lantern/BaseNode.test.js",
+    "core/lib/lantern/metrics/*.test.js",
+    "core/lib/lantern/PageDependencyGraph.test.js",
+    "core/lib/lantern/simulation/ConnectionPool.test.js",
+    "core/lib/lantern/simulation/DNSCache.test.js",
+    "core/lib/lantern/simulation/NetworkAnalyzer.test.js",
+    "core/lib/lantern/simulation/Simulator.test.js",
+    // ------ done TODO
     "core/test/lib/lh-element-test.js",
     "core/test/lib/lighthouse-compatibility-test.js",
     "core/test/lib/manifest-parser-test.js",
@@ -109,6 +111,6 @@
     "core/test/computed/metrics/interactive-test.js",
     "core/test/computed/tbt-impact-tasks-test.js",
     "core/test/fixtures/config-plugins/lighthouse-plugin-simple/plugin-simple.js",
-    "core/test/lib/lantern/metrics/MetricTestUtils.js",
+    "core/lib/lantern/metrics/MetricTestUtils.js",
   ],
 }
